The daily classifier bands deadlocks as a rate through the card band's tiers, not any-deadlock-is-Critical (#3564)
* The daily classifier bands deadlocks as a rate through the card band's tiers, not any-deadlock-is-Critical (#3525)

The shared DailyHealthBandCalculator still shipped #3368's un-fixed twin:
signals.Deadlocks > 0 banded the whole day Critical. On the measured
43-server fleet that read 87.9% of 24-hour windows Critical, so ~7 of 8
Performance Calendar day cells painted red from deadlocks alone, and the
fleet sweep's variable 15-1440min span scaled its Critical rate ~6x from
the cadence knob.

DailyHealthSignals now carries the Window its counts cover (the
ServerHealthMetrics.DeadlockWindow discipline: default(TimeSpan) is
unusable, never a divisor), and Classify routes the Deadlocks signal
through ServerHealthClassifier.DeadlockSeverity with the store-backed
DeadlockRateThresholds (V120) via DailyHealthThresholds.DeadlockRates —
Critical/Warning fold into the day's tiers, and the sub-1h/undeclared
window arm falls to Warning-not-rate (zero stays out of the trigger).
DeadlockRatePerHour/DeadlockSeverity widen int->long for the day-scale
counts; every int caller converts implicitly.

Producers declare their windows: the calendar-day projections (Darling
health reader, viewer, Lite) pass 24h, the fleet sweep passes its own
span. The Darling day reads hoist one store-tiers read per range/sweep
(the fleet reader's own hot-swap argument), the sweep's Compose takes
the tiers as a required parameter, and its would-have-paged deadlock
family now fires only when the rate crossed the Critical tier, with
rate + tier + count as evidence. The shared deadlock reason/tooltip
line reports the rate beside the count ("120 deadlocks (5.0/hr)"), the
card reason's own disclosure rule, and the sweep verdict JSON carries
deadlock_rate_per_hour + window_minutes additively.

Tests: two-window proofs on the day classifier (same rate bands the same
over 1h and 24h; same count bands differently), the single-deadlock day
control, the sub-hour arms, tier-override pins, sweep ledger pins, a
whole-repo census that every production DailyHealthSignals bundle
declares its Window (with the calendar projections pinned to 24h and
the sweep pinned to its span), and the Lite live-DuckDB calendar now
proves the rate path end-to-end with a 480-deadlock storm day.

Fixes #3525


* Two live health-tool tests catch up to #3525's rate banding

DarlingMcpHealthToolsLivePostgresTests planted one deadlock and asserted the
day reads Critical - the any-deadlock-is-Critical semantics #3525 removes.
One deadlock across a 24h day is 0.04/hr, far below the measured tiers, so
the boundary test now pins the honest Healthy band plus the deadlock_count
evidence (its real subject is date-boundary visibility), and the planted-rows
test pins Warning (its 85% CPU sample trips the HighCpuWarningSamples tier)
plus the same evidence. The Critical-band rate path is covered by the PR's
own storm-day tests.


* The still-forming day clamps its window to elapsed time (#3525 review)

The calendar-day projections stamped Window = 24h on every row including
today's half-open bucket, so an active storm diluted against hours that
had not happened yet: 60 deadlocks in the last hour read 2.5/hr (Healthy)
instead of 60/hr (Critical) - a false-negative the pre-#3525 any-deadlock
trigger could not produce. The shared CalendarDayWindow helper now clamps
the current day to its elapsed portion against the read's own clock: the
anchored range reads (both SKUs) hand their resolved as_of window end so
a backdated read clamps against itself, the live calendar reads use the
wall clock, and a day minutes old falls to the sub-hour Warning-not-rate
arm rather than a fabricated multiplied rate. Finished days band over
their full 24 hours exactly as before, and the sweep path is untouched.
The window census now pins all three projections to the clamp helper so
a hand-rolled FromDays(1) cannot reintroduce the dilution, and the twin
band suites carry the storm proof at the review's own numbers.


---------

+++ PerformanceMonitor.Common/ServerHealthBands.cs
@@ -697,8 +697,11 @@ namespace PerformanceMonitor.Common
         /// <para>Public because both cards REPORT the rate beside the count: a card that bands on a figure
         /// it does not show leaves an operator reading "Deadlocks 3" against a Critical dot with no way to
         /// see which number crossed which tier.</para>
+        ///
+        /// <para>The count is a <c>long</c> because the daily classifier's signals carry day-scale
+        /// roll-ups as longs (#3525); every <c>int</c> caller widens implicitly.</para>
         /// </summary>
-        public static double? DeadlockRatePerHour(int deadlockCount, TimeSpan window) =>
+        public static double? DeadlockRatePerHour(long deadlockCount, TimeSpan window) =>
             window >= ServerHealthThresholds.DeadlockRateMinimumWindow
                 ? deadlockCount / window.TotalHours
                 : null;
@@ -743,7 +746,10 @@ namespace PerformanceMonitor.Common
         /// keeps reading healthy; that reasoning is untouched here. A PostgreSQL target has no
         /// SQL-Server-deadlock reading at any rate, so it bands off none.</para></summary>
         /// <param name="deadlockCount">Deadlocks counted in the window, or null where the engine has no
-        /// source behind the reading.</param>
+        /// source behind the reading. A <c>long</c> for <see cref="DeadlockRatePerHour"/>'s reason (#3525):
+        /// the shared daily classifier routes its day-scale <c>Deadlocks</c> roll-up through this same band,
+        /// so the calendar, <c>get_daily_summary</c>, the fleet sweep and the Overview card cannot disagree
+        /// about what the same count over the same window means.</param>
         /// <param name="window">How long that count covers. Required rather than defaulted, for the reason
         /// <see cref="CpuSeverity"/>'s source is: a caller that kept the old single-argument call would
         /// compile and silently band a bare count again, which is the entire defect.</param>
@@ -751,7 +757,7 @@ namespace PerformanceMonitor.Common
         /// would let a surface band on the shipped pair while <c>get_alert_settings</c> reported the
         /// store's.</param>
         public static HealthSeverity DeadlockSeverity(
-            int? deadlockCount,
+            long? deadlockCount,
             TimeSpan window,
             DeadlockRateThresholds thresholds)
         {
